+++ dymension/libs/kaspa/demo/relayer/src/x/mod.rs
@@ -1,1 +1,3 @@
 pub mod args;
+
+pub mod deposit;
