+++ dymension/libs/kaspa/lib/core/src/lib.rs
@@ -17,9 +17,6 @@ use kaspa_rpc_core::RpcScriptPublicKey;
 use kaspa_txscript::extract_script_pub_key_address;
 pub use secp256k1::Keypair as KaspaSecpKeypair;
 
-pub const ESCROW_ADDRESS: &'static str =
-    "kaspatest:qzwyrgapjnhtjqkxdrmp7fpm3yddw296v2ajv9nmgmw5k3z0r38guevxyk7j0";
-
 pub fn parse_hyperlane_message(m: &RawHyperlaneMessage) -> Result<HyperlaneMessage, anyhow::Error> {
     const MIN_EXPECTED_LENGTH: usize = 77;
